separate dev server from production server
